Updated to Azure.messaging.EventHubs
