+++ EventHubSampleData/EventHubSampleSendData/Properties/AssemblyInfo.cs
