+++ EventHubSampleData/EventHubSampleSendData/packages.config
@@ -1,0 +1,26 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Azure.Core" version="1.5.0" targetFramework="net461" />
+  <package id="Azure.Messaging.EventHubs" version="5.2.0" targetFramework="net461" />
+  <package id="Microsoft.Azure.Amqp" version="2.4.5" targetFramework="net461" />
+  <package id="Microsoft.Bcl.AsyncInterfaces" version="1.0.0" targetFramework="net461" />
+  <package id="System.Buffers" version="4.5.0" targetFramework="net461" />
+  <package id="System.Diagnostics.DiagnosticSource" version="4.6.0" targetFramework="net461" />
+  <package id="System.IdentityModel.Tokens.Jwt" version="4.0.4.403061554" targetFramework="net461" />
+  <package id="System.Memory" version="4.5.3" targetFramework="net461" />
+  <package id="System.Net.WebSockets" version="4.0.0" targetFramework="net461" />
+  <package id="System.Net.WebSockets.Client" version="4.0.2" targetFramework="net461" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net461" />
+  <package id="System.Reflection.TypeExtensions" version="4.5.1" targetFramework="net461" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.6.0" targetFramework="net461" />
+  <package id="System.Runtime.Serialization.Primitives" version="4.1.1" targetFramework="net461" />
+  <package id="System.Security.Cryptography.Algorithms" version="4.2.0" targetFramework="net461" />
+  <package id="System.Security.Cryptography.Encoding" version="4.0.0" targetFramework="net461" />
+  <package id="System.Security.Cryptography.Primitives" version="4.0.0" targetFramework="net461" />
+  <package id="System.Security.Cryptography.X509Certificates" version="4.1.0" targetFramework="net461" />
+  <package id="System.Text.Encodings.Web" version="4.6.0" targetFramework="net461" />
+  <package id="System.Text.Json" version="4.6.0" targetFramework="net461" />
+  <package id="System.Threading.Channels" version="4.6.0" targetFramework="net461" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.2" targetFramework="net461" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net461" />
+</packages>
